requirements.txt
selenium
pytest
pyyaml
openpyxl
